+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/tests/test_role.py
@@ -10,7 +10,7 @@ import tempfile
 import time
 import unittest
 
-from azure.cli.testsdk import LiveScenarioTest, ResourceGroupPreparer, KeyVaultPreparer
+from azure.cli.testsdk import ScenarioTest, LiveScenarioTest, ResourceGroupPreparer, KeyVaultPreparer
 
 
 class RbacSPSecretScenarioTest(LiveScenarioTest):
@@ -229,6 +229,23 @@ class RoleAssignmentScenarioTest(LiveScenarioTest):
             self.cmd('role assignment delete --assignee {upn} --role reader')
         finally:
             self.cmd('ad user delete --upn-or-object-id {upn}')
+
+
+class RoleAssignmentListScenarioTest(ScenarioTest):
+
+    @ResourceGroupPreparer(name_prefix='cli_test_assignments_for_coadmins')
+    def test_assignments_for_co_admins(self, resource_group):
+        from azure_devtools.scenario_tests import LargeResponseBodyProcessor
+        large_resp_body = next((r for r in self.recording_processors if isinstance(r, LargeResponseBodyProcessor)), None)
+        if large_resp_body:
+            large_resp_body._max_response_body = 1024
+        result = self.cmd('role assignment list --include-classic-administrator').get_output_in_json()
+        self.assertTrue([x for x in result if x['properties']['roleDefinitionName'] in ['CoAdministrator', 'AccountAdministrator']])
+        self.cmd('role assignment list -g {}'.format(resource_group), checks=[
+            self.check("length([])", 0)
+        ])
+        result = self.cmd('role assignment list -g {} --include-classic-administrator'.format(resource_group)).get_output_in_json()
+        self.assertTrue([x for x in result if x['properties']['roleDefinitionName'] in ['CoAdministrator', 'AccountAdministrator']])
 
 
 if __name__ == '__main__':