+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 0.1.21
 ++++++
 * `webapp/functionapp`: ensure list/show display correct set of apps
+* webapp: set WEBSITE_NODE_DEFAULT_VERSION in case where runtime is not set
 
 0.1.20
 ++++++